requirements.txt
pymysql==1.0.2
python-decouple==3.3
sqlalchemy==1.4.25
